+++ python/GafferUITest/NumericPlugValueWidgetTest.py
@@ -90,6 +90,30 @@ class NumericPlugValueWidgetTest( GafferUITest.TestCase ) :
 		Gaffer.MetadataAlgo.setReadOnly( n["user"]["i1"], True )
 		self.assertFalse( w.numericWidget().getEditable() )
 
+	def testMixedOrInvalidValuesPreservesExisting( self ) :
+
+		n = Gaffer.Node()
+		n["user"]["i1"] = Gaffer.IntPlug()
+		n["user"]["i2"] = Gaffer.IntPlug()
+		n["user"]["i1"].setValue( 1 )
+		n["user"]["i2"].setValue( 2 )
+
+		w = GafferUI.NumericPlugValueWidget( n["user"]["i1"] )
+		self.assertEqual( w.numericWidget().getValue(), 1 )
+
+		w.numericWidget().setText( "" )
+		w.numericWidget()._qtWidget().editingFinished.emit()
+
+		self.assertEqual( n["user"]["i1"].getValue(), 1 )
+
+		w = GafferUI.NumericPlugValueWidget( n["user"].children() )
+		self.assertEqual( w.numericWidget().getText(), "" )
+
+		w.numericWidget()._qtWidget().editingFinished.emit()
+
+		self.assertEqual( n["user"]["i1"].getValue(), 1 )
+		self.assertEqual( n["user"]["i2"].getValue(), 2 )
+
 	def testFixedCharacterWidth( self ) :
 
 		n = Gaffer.Node()