+++ Changes.md
@@ -5,6 +5,12 @@ Fixes
 -----
 
 - Viewer : Fixed bug that caused the Inspector to grow very wide when inspecting long strings.
+- NumericPlugValueWidget : Fixed bug that caused plug values to be set to 0 if editing completed with invalid text.
+
+API
+---
+
+- NumericWidget : Fixed bug that caused `editingFinished` to be called with the wrong `reason` when the widget was left with an invalid value. `Invalid` is now passed in these cases.
 
 0.58.4.0 (relative to 0.58.3.2)
 ========